feature-002: feedback 보내기 컴포넌트로 분리
