+++ src/components/layout/footer/Footer.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react';
 import SvgIcons from '../../SvgIcons';
 import * as FooterStyle from '@/styles/layout/Footer.style';
 import AboutViNO from './AboutViNO';
+import SendEmail from './SendEmail';
 
 const Footer = () => {
   const [feedback, setFeedback] = useState<string>('');
@@ -11,25 +12,10 @@ const Footer = () => {
     ['이용약관', '개인정보처리방침', '문의하기'],
   ];
 
-  const handleInputFeedback = (e: React.ChangeEvent<HTMLInputElement>) =>
-    setFeedback(e.target.value);
-
   return (
     <FooterStyle.Container>
       <SvgIcons.Logo width={64.55} height={20} footer />
-      <FooterStyle.SendEmailWrap>
-        <FooterStyle.SendEmailImage
-          src="/src/assets/mail.png"
-          alt="메일 보내기"
-        />
-        <FooterStyle.SendEmailInput
-          type="text"
-          placeholder="서비스를 이용하면서 불편하거나, 좋은 피드백이 있다면 보내주세요!"
-          value={feedback}
-          onChange={handleInputFeedback}
-        />
-        <FooterStyle.SendEmailButton>보내기</FooterStyle.SendEmailButton>
-      </FooterStyle.SendEmailWrap>
+      <SendEmail feedback={feedback} setFeedback={setFeedback} />
       <AboutViNO aboutViNOs={aboutViNOs} />
       <FooterStyle.CopyRight>
         © 2024 Vino. All rights reserved.
